Add a minimum group size field
Adds a minimum group size field. If a group is below that size, then projects above the threshold are random selected and it checks the next-highest preference for every member in that group. If the next-highest is below the minimum, they are moved into the below-minimum group and removed from their current group. If no one in that project has the next-highest preference that is below the minimum, it selects another project randomly, but without repeats. If all projects are checked, it repeats and checks the preferences again from random projects until the groups all have the minimum threshold.